+++ examples/Gjs/gnome-shell-hello-world/.ts-for-girrc.js
@@ -23,6 +23,6 @@ export default {
     ],
     pretty: true,
     buildType: "types",
-    useNamespace: true,
+    noNamespace: false,
     gnomeShellTypes: true,
 }