+++ packages/cli/src/config.ts
@@ -13,11 +13,11 @@ import { writeFile } from 'fs/promises'
 import { Logger } from './logger.js'
 import { APP_NAME, APP_USAGE } from './constants.js'
 import {
-    WARN_USE_NAMESPACE_ON_TYPES,
-    WARN_USE_NAMESPACE_ON_ESM,
+    WARN_NO_NAMESPACE_ON_TYPES,
+    WARN_NO_NAMESPACE_ON_ESM,
+    ERROR_CONFIG_EXTENSION_UNSUPPORTED,
     WARN_USE_ESM_FOR_ALIAS,
     WARN_USE_GJS_FOR_ALIAS,
-    ERROR_CONFIG_EXTENSION_UNSUPPORTED,
 } from './messages.js'
 
 export class Config {
@@ -29,7 +29,7 @@ export class Config {
      * Default cli flag and argument values
      */
     static defaults = {
-        environments: ['gjs', 'node'],
+        environments: ['gjs'],
         pretty: false,
         print: false,
         configName: '.ts-for-girrc.js',
@@ -39,9 +39,9 @@ export class Config {
         ignore: [],
         verbose: false,
         ignoreVersionConflicts: false,
-        useNamespace: false,
+        noNamespace: false,
         buildType: 'lib',
-        moduleType: 'commonjs',
+        moduleType: 'esm',
         noComments: false,
         noDebugComments: false,
         noCheck: false,
@@ -144,11 +144,11 @@ export class Config {
             default: Config.defaults.configName,
             normalize: true,
         },
-        useNamespace: {
+        noNamespace: {
             type: 'boolean',
             alias: 'd',
-            description: 'Export all symbols for each module as a namespace',
-            default: Config.defaults.useNamespace,
+            description: 'Do not export all symbols for each module as a namespace',
+            default: Config.defaults.noNamespace,
             normalize: true,
         },
         noComments: {
@@ -213,7 +213,7 @@ export class Config {
         ignoreVersionConflicts: this.options.ignoreVersionConflicts,
         print: this.options.print,
         configName: this.options.configName,
-        useNamespace: this.options.useNamespace,
+        noNamespace: this.options.noNamespace,
         noComments: this.options.noComments,
         noDebugComments: this.options.noDebugComments,
         noCheck: this.options.noCheck,
@@ -303,7 +303,7 @@ export class Config {
             verbose: config.verbose,
             buildType: config.buildType,
             moduleType: config.moduleType,
-            useNamespace: config.useNamespace,
+            noNamespace: config.noNamespace,
             noComments: config.noComments,
             noDebugComments: config.noDebugComments,
             noCheck: config.noCheck,
@@ -344,16 +344,16 @@ export class Config {
 
     public static async validate(config: UserConfig): Promise<UserConfig> {
         if (config.buildType === 'types') {
-            if (config.useNamespace !== true) {
-                Logger.warn(WARN_USE_NAMESPACE_ON_TYPES)
-                config.useNamespace = true
+            if (config.noNamespace !== false) {
+                Logger.warn(WARN_NO_NAMESPACE_ON_TYPES)
+                config.noNamespace = false
             }
         }
 
         if (config.moduleType === 'esm') {
-            if (config.useNamespace !== true) {
-                Logger.warn(WARN_USE_NAMESPACE_ON_ESM)
-                config.useNamespace = true
+            if (config.noNamespace !== false) {
+                Logger.warn(WARN_NO_NAMESPACE_ON_ESM)
+                config.noNamespace = false
             }
         }
 
@@ -369,9 +369,9 @@ export class Config {
         }
 
         if (config.moduleType === 'esm') {
-            if (config.useNamespace !== true) {
-                Logger.warn(WARN_USE_NAMESPACE_ON_ESM)
-                config.useNamespace = true
+            if (config.noNamespace !== false) {
+                Logger.warn(WARN_NO_NAMESPACE_ON_ESM)
+                config.noNamespace = false
             }
         }
 
@@ -400,7 +400,7 @@ export class Config {
             girDirectories: options.girDirectories,
             ignore: options.ignore,
             modules: options.modules,
-            useNamespace: options.useNamespace,
+            noNamespace: options.noNamespace,
             noComments: options.noComments,
             noDebugComments: options.noDebugComments,
             noCheck: options.noCheck,
@@ -464,12 +464,12 @@ export class Config {
             ) {
                 config.modules = configFile.modules
             }
-            // useNamespace
+            // noNamespace
             if (
-                config.useNamespace === Config.options.useNamespace.default &&
-                typeof configFile.useNamespace === 'boolean'
+                config.noNamespace === Config.options.noNamespace.default &&
+                typeof configFile.noNamespace === 'boolean'
             ) {
-                config.useNamespace = configFile.useNamespace
+                config.noNamespace = configFile.noNamespace
             }
             // noComments
             if (config.noComments === Config.options.noComments.default && typeof configFile.noComments === 'boolean') {