+++ packages/cli/templates/Gjs/ui/animation.d.ts
@@ -4,14 +4,14 @@
 <% const Gio = dep.find('Gio') %>
 <% const St = dep.find('St') %>
 
-<%_ if(useNamespace){ _%>
-    import type GLib from '../<%= glib.packageName %>.js';
-    import type Gio from '../<%= Gio.packageName %>.js';
-    import type St from '../<%= St.packageName %>.js';
-<%_ } else { _%>
+<%_ if(noNamespace){ _%>
     import type * as GLib from '../<%= glib.packageName %>.js';
     import type * as Gio from '../<%= Gio.packageName %>.js';
     import type * as St from '../<%= St.packageName %>.js';
+<%_ } else { _%>
+    import type GLib from '../<%= glib.packageName %>.js';
+    import type Gio from '../<%= Gio.packageName %>.js';
+    import type St from '../<%= St.packageName %>.js';
 <%_ } _%>
 
 export class Animation extends St.Bin {