+++ packages/cli/templates/Gjs/ui/main.d.ts
@@ -1,13 +1,13 @@
 <% const Gio = dep.find('Gio') %>
 
-<%_ if(useNamespace){ _%>
-    import type Gio from '../<%= Gio.packageName %>.js';
-<%_ } else { _%>
+<%_ if(noNamespace){ _%>
     import type * as Gio from '../<%= Gio.packageName %>.js';
+<%_ } else { _%>
+    import type Gio from '../<%= Gio.packageName %>.js';
 <%_ } _%>
 
 import { ComponentManager } from './components/__init__.js';
-// const AccessDialog = imports.ui.accessDialog;
+import { AccessDialogDBus } from './accessDialog.js';
 // const AudioDeviceSelection = imports.ui.audioDeviceSelection;
 // const Components = imports.ui.components;
 // const CtrlAltTab = imports.ui.ctrlAltTab;
@@ -68,7 +68,7 @@ export declare const shellMountOpDBusService: any;
 
 export declare const shellDBusService: any;
 
-export declare const shellAccessDialogDBusService: any;
+export declare const shellAccessDialogDBusService: AccessDialogDBus;
 
 export declare const shellAudioSelectionDBusService: any;