+++ packages/cli/templates/Gjs/ui/modalDialog.d.ts
@@ -3,12 +3,12 @@
 <% const St = dep.find('St') %>
 <% const shell = dep.find('Shell') %>
 
-<%_ if(useNamespace){ _%>
-    import type St from '../<%= St.packageName %>.js';
-    import type Shell from '../<%= shell.packageName %>.js';
-<%_ } else { _%>
+<%_ if(noNamespace){ _%>
     import type * as St from '../<%= St.packageName %>.js';
     import type * as Shell from '../<%= shell.packageName %>.js';
+<%_ } else { _%>
+    import type St from '../<%= St.packageName %>.js';
+    import type Shell from '../<%= shell.packageName %>.js';
 <%_ } _%>
 
 import type { MonitorConstraint } from './layout.js';